+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support local assets for 3D puck and `ModelLayer`. To use a local assets, please specify it with `asset://` scheme in the uri.
 * Fix map view crashing upon host activity destruction when using a cached Flutter engine.
 * Fix a rare crash happening when map widget is being disposed.
+* `MapDebugOptions` got superseeded by `MapWidgetDebugOptions`, expanding existing debug options with the new `light`, `camera`, `padding` debug options in addition to the new Android-specific options: `layers2DWireframe`, `layers3DWireframe`.
 
 ### 2.1.0